refactor: move contents from the src/models/apiModels to parent directory
