+++ src/contentScripts/views/Home/components/Following.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { Ref } from 'vue'
-import type { DataItem as MomentItem, MomentResult } from '~/models/apiModels/moment/moment'
+import type { DataItem as MomentItem, MomentResult } from '~/models/moment/moment'
 import emitter from '~/utils/mitt'
 
 const videoList = reactive<MomentItem[]>([])
@@ -124,4 +124,3 @@ function jumpToLoginPage() {
 
 <style lang="scss" scoped>
 </style>
-~/models/apiModels/moment/momentResult
