+++ src/models/anime/popular.ts
