+++ src/models/anime/timeTable.ts
